Add login screen states switching via navigator.
